split source finding code out so it can be run/tested separately
